# swagger.yaml
openapi: 3.0.3
info:
  title: Smart Vending System API
  description: |
    Flask backend for smart vending: catalog, Omise checkout, members (OTP/JWT), orders, promotions, and admin CRUD under `/api/admin`.
    Interactive docs (Flasgger) are also served by the server at `/apidocs` when the app is running.

    **Socket.IO (same host/port as HTTP, e.g. http://localhost:8000):**
    - Pi agent: auth with `{ "machine_code": "<code>", "token": "<secret>" }` (bcrypt on server). Server emits `job.start` after checkout; agent replies via machine_event RPC (persisted and `job_event_broadcast` to kiosk).
    - Kiosk (machine-ui): auth `{ "role": "kiosk", "kiosk_secret": "..." }` — required in production. Listens for `job_event_broadcast`.
    - Admin dashboard: `{ "role": "admin" }` or `{ "admin_token": "..." }` when `ADMIN_SOCKET_SECRET` is set. Receives `dashboard_update` on machine events.

    Checkout dispatches hardware jobs via Socket.IO `emit_job_start` only (no REST `/api/machines/events`). Manual agent test: POST http://localhost:5000/jobs/start.

    **Docker Compose — default published ports on the host:**

    | Service | Host URL |
    |---------|----------|
    | MySQL (`db`) | `localhost:3307` |
    | API + Socket.IO (`server`) | `http://localhost:8000` |
    | Hardware agent (`client`) | `http://localhost:5000` |
    | machine-ui | `http://localhost:3000` |
    | admin-ui | `http://localhost:3001` |
    | web-ui (member app) | `http://localhost:3002` |
...
servers:
...
paths:
...
  /api/admin/auth/login:
    post:
      tags:
      - Admin Auth
      summary: Authenticate admin with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              - password
              properties:
                email:
                  type: string
                  format: email
                  example: admin@modpao.com
                password:
                  type: string
                  format: password
                  example: admin1234
      responses:
        '200':
          description: Successfully authenticated or redirects to complete registration
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT Bearer Access Token (if active)
                  user:
                    $ref: '#/components/schemas/AdminProfile'
                  needs_registration:
                    type: boolean
                    description: True if admin has not completed onboarding
                  registration_token:
                    type: string
                    description: One-time token used to invoke /auth/register
                  email:
                    type: string
        '400':
          description: Email and password are required
        '401':
          description: Invalid credentials
  /api/admin/auth/register:
    post:
      tags:
      - Admin Auth
      summary: Complete onboarding for invited admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - registration_token
              - new_password
              properties:
                registration_token:
                  type: string
                new_password:
                  type: string
                  format: password
                first_name:
                  type: string
                last_name:
                  type: string
                position:
                  type: string
                phone:
                  type: string
      responses:
        '200':
          description: Registration completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/AdminProfile'
        '400':
          description: Missing parameters or account already active
        '401':
          description: Invalid/expired registration token
        '404':
          description: Admin user not found
  /api/admin/auth/invite:
    post:
      tags:
      - Admin Auth
      summary: Invite a new administrator (Superadmin only)
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              - temp_password
              properties:
                email:
                  type: string
                  format: email
                temp_password:
                  type: string
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/AdminProfile'
                  invite_link:
                    type: string
                  message:
                    type: string
        '400':
          description: Email and temporary password required
        '401':
          description: Unauthorized
        '403':
          description: Requires Superadmin privileges
        '409':
          description: Email already exists
  /api/admin/auth/revoke/{admin_id}:
    delete:
      tags:
      - Admin Auth
      summary: Revoke admin access and delete account (Superadmin only)
      security:
      - bearerAuth: []
      parameters:
      - name: admin_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Access revoked successfully
        '400':
          description: Cannot revoke your own account
        '401':
          description: Unauthorized
        '403':
          description: Requires Superadmin privileges
        '404':
          description: Admin not found
  /api/admin/auth/me:
    get:
      tags:
      - Admin Auth
      summary: Fetch current logged-in admin profile
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Active profile details
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/AdminProfile'
        '401':
          description: Unauthorized
        '404':
          description: Admin not found
  /api/admin/auth/admins:
    get:
      tags:
      - Admin Auth
      summary: List all administrator accounts (Superadmin only)
      security:
      - bearerAuth: []
      responses:
        '200':
          description: List of admins
          content:
            application/json:
              schema:
                type: object
                properties:
                  admins:
                    type: array
                    items:
                      allOf:
                      - $ref: '#/components/schemas/AdminProfile'
                      - type: object
                        properties:
                          status:
                            type: string
                            example: Active
        '401':
          description: Unauthorized
        '403':
          description: Requires Superadmin privileges
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer JWT token (Member JWT or Admin JWT based on endpoint)
  schemas:
...
    AdminProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        position:
          type: string
        phone:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          nullable: true
...